# doc/api.yaml
openapi: 3.0.0
info:
  title: Home Library Service
  description: Home music library service
  version: 1.0.0

servers:
  - url: http://localhost:4000

components:
  schemas:
    RequestBodyValidationError:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        message:
          type: array
          items:
            type: string
          example: ['field should not be empty']
        error:
          type: string
          example: 'Bad Request'
    UnprocessableEntityError:
      type: object
      properties:
        statusCode:
          type: number
          example: 422
        message:
          type: string
          example: 'Unprocessable Entity'
    NotFoundError:
      type: object
      properties:
        statusCode:
          type: number
          example: 404
        message:
          type: string
          example: 'Not Found'
    ValidationError:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        message:
          type: string
          example: 'Validation failed (uuid  is expected)'
        error:
          type: string
          example: 'Bad Request'
    ForbiddenError:
      type: object
      properties:
        statusCode:
          type: number
          example: 403
        message:
          type: string
          example: 'Forbidden'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        login:
          type: string
          example: 'TestUser'
        version:
          type: number
          example: 1
        createAt:
          type: number
          example: 1655000000
        updateAt:
          type: number
          example: 1655000000
      required:
        - id
        - login
    Artist:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Freddie Mercury'
        grammy:
          type: boolean
          example: false
      required:
        - id
        - name
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Innuendo'
        year:
          type: integer
          example: 1991
        artistId:
          type: string
          format: uuid
          nullable: true
      required:
        - id
        - name
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: The Show Must Go On
        artistId:
          type: string
          format: uuid
          nullable: true
        albumId:
          type: string
          format: uuid
          nullable: true
        duration:
          type: integer
          description: In seconds
          example: 262
      required:
        - id
        - name
        - duration
    Favorites:
      type: object
      properties:
        artists:
          type: array
          items:
            $ref: '#/components/schemas/Artist'
        albums:
          type: array
          items:
            $ref: '#/components/schemas/Album'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []

paths:
  /login:
    post:
      tags:
        - Login
      security: []
      summary: Login
      description: Logins a user and returns a JWT-token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                login:
                  type: string
                  description: Username
                password:
                  type: string
                  description: Password
              required:
                - user
                - login
      responses:
        200:
          description: Successful login.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                description: JWT Token
        403:
          description: Incorrect login or password
  /signup:
    post:
      tags:
        - Signup
      security: []
      summary: Signup
      description: Signup a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                login:
                  type: string
                  minLength: 3
                  maxLength: 255
                  description: Username
                password:
                  type: string
                  format: password
                  pattern: ^[a-zA-Z0-9]{3,30}
                  description: Password
              required:
                - login
                - password
      responses:
        204:
          description: Successful signup
        400:
          description: Bad request
        409:
          description: Conflict. Login already exists
  /user:
    get:
      tags:
        - Users
      summary: Get all users
      description: Gets all users
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        401:
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Users
      summary: Create user
      description: Creates a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                login:
                  type: string
                  description: The user's login
                password:
                  type: string
                  description: The user's password
              required:
                - login
                - password
      responses:
        201:
          description: The user has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Bad request. body does not contain required fields
          content:
            application/json:
              schema:
                $ref: '#components/schemas/RequestBodyValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /user/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          pattern: uuid
    get:
      tags:
        - Users
      summary: Get single user by id
      description: Get single user by id
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: Bad request. userId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'

    put:
      tags:
        - Users
      summary: Update a user's password
      description: Updates a user's password by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                oldPassword:
                  type: string
                  description: The user's old password
                newPassword:
                  type: string
                  description: The user's new password
              required:
                - oldPassword
                - newPassword
      responses:
        200:
          description: The user has been updated.
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  id:
                    type: string
                    format: uuid
                  login:
                    type: string
                    example: TestUser
                  version:
                    type: integer
                    example: 2
                  createAt:
                    type: integer
                    example: 1655000000
                  updateAt:
                    type: integer
                    example: 1655999999
        400:
          description: Bad request. userId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          description: oldPassword is wrong
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ForbiddenError'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Deletes user by ID.
      responses:
        204:
          description: The user has been deleted
        400:
          description: Bad request. userId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'
  /track:
    get:
      tags:
        - Track
      summary: Get tracks list
      description: Gets all library tracks list
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
        401:
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Track
      summary: Add new track
      description: Add new track information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                artistId:
                  type: string
                  format: uuid
                albumId:
                  type: string
                  format: uuid
                duration:
                  type: integer
                  description: In seconds
              required:
                - name
                - duration
      responses:
        201:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        400:
          description: Bad request. body does not contain required fields
          content:
            application/json:
              schema:
                $ref: '#components/schemas/RequestBodyValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /track/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Track
      summary: Get single track by id
      description: Gets single track by id
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        400:
          description: Bad request. trackId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Track was not found.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'
    put:
      tags:
        - Track
      summary: Update track information
      description: Update library track information by UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                  example: Bohemian Rhapsody
                artistId:
                  type: string
                  format: uuid
                  nullable: true
                duration:
                  type: integer
                  description: In seconds
                  example: 355
                albumId:
                  type: string
                  format: uuid
                  nullable: true
              required:
                - name
                - duration
      responses:
        200:
          description: The track has been updated.
          content:
            application/json:
              schema:
                type: object
                title: example
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                    example: Bohemian Rhapsody
                  artistId:
                    type: string
                    format: uuid
                    nullable: true
                  duration:
                    type: integer
                    description: In seconds
                    example: 355
                  albumId:
                    type: string
                    format: uuid
                    nullable: true
        400:
          description: Bad request. trackId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Track was not found.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'
    delete:
      tags:
        - Track
      summary: Delete track
      description: Delete track from library
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request. trackId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Track was not found.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'
  /album:
    get:
      tags:
        - Album
      summary: Get albums list
      description: Gets all library albums list
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
        401:
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Album
      summary: Add new album
      description: Add new album information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                year:
                  type: integer
                artistId:
                  type: string
                  format: uuid
                  nullable: true
              required:
                - name
                - year
      responses:
        201:
          description: Album is created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        400:
          description: Bad request. body does not contain required fields
          content:
            application/json:
              schema:
                $ref: '#components/schemas/RequestBodyValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /album/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Album
      summary: Get single album by id
      description: Gets single album by id
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        400:
          description: Bad request. albumId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'
    put:
      tags:
        - Album
      summary: Update album information
      description: Update library album information by UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: example
              properties:
                name:
                  type: string
                year:
                  type: integer
                artistId:
                  type: string
                  format: uuid
                  nullable: true
              required:
                - name
                - year
      responses:
        200:
          description: The album has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        400:
          description: Bad request. albumId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'
    delete:
      tags:
        - Album
      summary: Delete album
      description: Delete album from library
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request. albumId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'
  /artist:
    get:
      tags:
        - Artist
      summary: Get all artists
      description: Gets all artists
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
        401:
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Artist
      summary: Add new artist
      description: Add new artist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                grammy:
                  type: boolean
              required:
                - name
                - grammy
      responses:
        201:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        400:
          description: Bad request. body does not contain required fields
          content:
            application/json:
              schema:
                $ref: '#components/schemas/RequestBodyValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /artist/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Artist
      summary: Get single artist by id
      description: Get single artist by id
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        400:
          description: Bad request. artistId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Artist was not found.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'
    put:
      tags:
        - Artist
      summary: Update artist information
      description: Update artist information by UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                grammy:
                  type: boolean
              required:
                - name
                - grammy
      responses:
        200:
          description: The artist has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        400:
          description: Bad request. artistId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Artist was not found.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'
    delete:
      tags:
        - Artist
      summary: Delete artist
      description: Delete artist from library
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request. artistId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Artist was not found.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'
  /favs:
    get:
      tags:
        - Favorites
      summary: Get all favorites
      description: Gets all favorites movies, tracks and books
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorites'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /favs/track/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags:
        - Favorites
      summary: Add track to the favorites
      description: Add track to the favorites
      responses:
        201:
          description: Added successfully
        400:
          description: Bad request. trackId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        422:
          description: Track with id doesn't exist.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/UnprocessableEntityError'
    delete:
      tags:
        - Favorites
      summary: Delete track from favorites
      description: Delete track from favorites
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request. trackId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Track was not found.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'
  /favs/album/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags:
        - Favorites
      summary: Add album to the favorites
      description: Add album to the favorites
      responses:
        201:
          description: Added successfully
        400:
          description: Bad request. albumId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        422:
          description: Album with id doesn't exist.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/UnprocessableEntityError'
    delete:
      tags:
        - Favorites
      summary: Delete album from favorites
      description: Delete album from favorites
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request. albumId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Album was not found.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'
  /favs/artist/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags:
        - Favorites
      summary: Add artist to the favorites
      description: Add artist to the favorites
      responses:
        201:
          description: Added successfully
        400:
          description: Bad request. artistId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        422:
          description: Artist with id doesn't exist.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/UnprocessableEntityError'
    delete:
      tags:
        - Favorites
      summary: Delete artist from favorites
      description: Delete artist from favorites
      responses:
        204:
          description: Deleted successfully
        400:
          description: Bad request. artistId is invalid (not uuid)
          content:
            application/json:
              schema:
                $ref: '#components/schemas/ValidationError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          description: Artist was not found.
          content:
            application/json:
              schema:
                $ref: '#components/schemas/NotFoundError'